开始开发表单部分

+++ java/form-service/src/main/java/com/plumdo/form/repository/FormDefinitionRepository.java
@@ -3,9 +3,10 @@ package com.plumdo.form.repository;
 
 import org.springframework.data.jpa.repository.Query;
 
-import com.plumdo.form.domian.FormDefinition;
+import com.plumdo.common.repository.BaseRepository;
+import com.plumdo.form.domain.FormDefinition;
 
-public interface FormDefinitionRepository extends BaseRepository<FormDefinition, Long> {
+public interface FormDefinitionRepository extends BaseRepository<FormDefinition, Integer> {
 	
 	@Query("select f from FormDefinition f where f.key = ?1 and (tenantId = ''  or tenantId is null) "
 			+ " and version = (select max(version) from FormDefinition where key = ?1 and (tenantId = ''  or tenantId is null))")
